# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/shift-types/{id}": {
      "get": {
        "operationId": "ShiftTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["ShiftTypes"]
      },
      "patch": {
        "operationId": "ShiftTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateShiftTypeDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["ShiftTypes"]
      },
      "delete": {
        "operationId": "ShiftTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["ShiftTypes"]
      }
    },
...
    "/api/payment-rules/shift-type/{shiftTypeId}": {
      "get": {
        "description": "Retrieves all payment rules for a specific shift type",
        "operationId": "PaymentRulesController_findByShiftType",
        "parameters": [
          {
            "name": "shiftTypeId",
            "required": true,
            "in": "path",
            "description": "The shift type ID",
            "schema": { "type": "string" }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": { "type": "boolean" }
          },
          {
            "name": "effectiveDate",
            "required": false,
            "in": "query",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "paymentType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all payment rules for the given shift type.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentRuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find payment rules by shift type",
        "tags": ["Payment Rules"]
      }
    },
    "/api/shift-type-premiums": {
      "post": {
        "description": "Creates a premium payment rule for a specific shift type and compensation rate",
        "operationId": "ShiftTypePremiumsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShiftTypePremiumDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The shift premium has been successfully created.",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new shift type premium",
        "tags": ["Shift Type Premiums"]
      },
      "get": {
        "description": "Retrieves all shift type premiums with optional filtering",
        "operationId": "ShiftTypePremiumsController_findAll",
        "parameters": [
          {
            "name": "shiftTypeId",
            "required": false,
            "in": "query",
            "description": "Filter by shift type ID",
            "schema": {
              "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "type": "string"
            }
          },
          {
            "name": "compensationRateId",
            "required": false,
            "in": "query",
            "description": "Filter by compensation rate ID",
            "schema": {
              "example": "3fa85f64-5717-4562-b3fc-2c963f66afa7",
              "type": "string"
            }
          },
          {
            "name": "effectiveDate",
            "required": false,
            "in": "query",
            "description": "Filter by effective date - returns premiums effective at the specified date",
            "schema": {
              "format": "date-time",
              "example": "2025-06-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter active premiums only (premiums with no end date or end date in the future)",
            "schema": { "example": true, "type": "boolean" }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip for pagination",
            "schema": { "minimum": 0, "example": 0, "type": "number" }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take for pagination",
            "schema": { "minimum": 1, "example": 10, "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all shift premiums that match the query.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "type": "object" } }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find all shift premiums",
        "tags": ["Shift Type Premiums"]
      }
    },
    "/api/shift-type-premiums/{id}": {
      "get": {
        "description": "Retrieves a specific shift premium by its unique identifier",
        "operationId": "ShiftTypePremiumsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The shift premium ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the shift premium with the given ID.",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": { "description": "Shift premium not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find one shift premium by ID",
        "tags": ["Shift Type Premiums"]
      },
      "patch": {
        "description": "Updates an existing shift premium with the provided data",
        "operationId": "ShiftTypePremiumsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The shift premium ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShiftTypePremiumDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The shift premium has been successfully updated.",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": { "description": "Shift premium not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a shift premium",
        "tags": ["Shift Type Premiums"]
      },
      "delete": {
        "description": "Removes a shift premium from the system",
        "operationId": "ShiftTypePremiumsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The shift premium ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The shift premium has been successfully deleted.",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": { "description": "Shift premium not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Delete a shift premium",
        "tags": ["Shift Type Premiums"]
      }
    },
    "/api/shift-type-premiums/applicable/{staffProfileId}/{departmentId}/{shiftTypeId}": {
      "get": {
        "description": "Finds all applicable premiums for a staff member working a specific shift type in a department",
        "operationId": "ShiftTypePremiumsController_findApplicablePremiums",
        "parameters": [
          {
            "name": "staffProfileId",
            "required": true,
            "in": "path",
            "description": "The staff profile ID",
            "schema": { "type": "string" }
          },
          {
            "name": "departmentId",
            "required": true,
            "in": "path",
            "description": "The department ID",
            "schema": { "type": "string" }
          },
          {
            "name": "shiftTypeId",
            "required": true,
            "in": "path",
            "description": "The shift type ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all applicable shift premiums.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "type": "object" } }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find applicable premiums",
        "tags": ["Shift Type Premiums"]
      }
    },
...
  },
  "info": {
    "title": "Multi-Sector Management API",
    "description": "API for managing organizations across multiple sectors",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "UpdateShiftTypeDto": { "type": "object", "properties": {} },
...
      "PaymentRuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa9"
          },
          "shiftTypeId": {
            "type": "string",
            "description": "The ID of the shift type for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "roleId": {
            "type": "string",
            "description": "The ID of the role for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa7"
          },
          "paymentType": {
            "type": "string",
            "description": "The type of payment",
            "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
            "example": "HOURLY"
          },
          "baseRate": {
            "type": "number",
            "description": "The base payment rate",
            "example": 25.5
          },
          "specialtyBonus": {
            "type": "number",
            "description": "Additional bonus for specialized skills",
            "example": 5
          },
          "experienceMultiplier": {
            "type": "number",
            "description": "Multiplier applied based on experience level",
            "example": 1.2
          },
          "effectiveDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule becomes effective",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule expires",
            "example": "2025-12-31T00:00:00.000Z",
            "nullable": true
          },
          "organizationId": {
            "type": "string",
            "description": "The ID of the organization this payment rule belongs to",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa8"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the payment rule was created",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the payment rule was last updated",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "shiftType": {
            "type": "object",
            "description": "The related shift type details"
          },
          "role": {
            "type": "object",
            "description": "The related role details"
          }
        },
        "required": [
          "id",
          "shiftTypeId",
          "roleId",
          "paymentType",
          "baseRate",
          "specialtyBonus",
          "experienceMultiplier",
          "effectiveDate",
          "organizationId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateShiftTypePremiumDto": {
        "type": "object",
        "properties": {
          "shiftTypeId": {
            "type": "string",
            "description": "The ID of the shift type this premium applies to",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "compensationRateId": {
            "type": "string",
            "description": "The ID of the compensation rate this premium applies to",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa7"
          },
          "isPremiumPercentage": {
            "type": "boolean",
            "description": "Whether the premium is calculated as a percentage (true) or fixed amount (false)",
            "example": true
          },
          "premiumValue": {
            "type": "number",
            "description": "The premium value (percentage or fixed amount)",
            "example": 0.25,
            "minimum": 0
          },
          "effectiveDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this shift premium becomes effective",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this shift premium expires (optional)",
            "example": "2025-12-31T00:00:00.000Z"
          }
        },
        "required": [
          "shiftTypeId",
          "compensationRateId",
          "isPremiumPercentage",
          "premiumValue",
          "effectiveDate"
        ]
      },
      "UpdateShiftTypePremiumDto": {
        "type": "object",
        "properties": {
          "isPremiumPercentage": {
            "type": "boolean",
            "description": "Whether the premium is calculated as a percentage (true) or fixed amount (false)",
            "example": true
          },
          "premiumValue": {
            "type": "number",
            "description": "The premium value (percentage or fixed amount)",
            "example": 0.25,
            "minimum": 0
          },
          "effectiveDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this shift premium becomes effective",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this shift premium expires",
            "example": "2025-12-31T00:00:00.000Z",
            "nullable": true
          }
        }
      },
...
    }
  }
}